Correct how CURRENT_PASSWORD is treated.
The VMS connector had assumed that when CURRENT_PASSWORD was specified, the configuration was that of the user being changed. It no longer assumes that, and instead, creates a separate connection to the VMS box, logging in as the user whose password is being changed, and executes a SET PASSWORD (or changes the password as part of the login sequence if it is already expired).

It's important to change password this way, because administrative password changes are subject to fewer checks than user password changes.

Since this change made the script considerably more complex, I've exported the various groovy scripts from the VmsConnector into separate files for ease of maintenance.

git-svn-id: https://svn.forgerock.org/openicf/trunk@292 05b3e5af-d696-470f-a577-fd7599f74d3c