Switch to just using name as Uid.
This allows the reporting of OWNER (for acount, group, and connection) to not require determining the object class.

LdapUtil needs to be updated to add in the Ldap portion to Uids when makiing Ldap calls, and may need to determine object class of OWNER before doing a set, but I can't determine this until I have a running Ldap system.

git-svn-id: https://svn.forgerock.org/openicf/trunk/projects/bundles/racf@4562 05b3e5af-d696-470f-a577-fd7599f74d3c